feat: move checkout logic to custom hook
